# server/openapi.yaml
openapi: 3.0.3
info:
  title: 蔓蔓点菜 API
  description: |
    蔓蔓点菜微信小程序后端 REST API。

    **认证方式**：所有标注 requireAuth 的端点需在请求头传递 `X-WX-OPENID: <用户openid>`。
    管理端点需额外传递 `X-WX-ADMIN: true`。

    **通用错误 envelope**：
    ```json
    { "error": { "code": "ERROR_CODE", "message": "中文描述" }, "requestId": "uuid" }
    ```
  version: 1.0.0
  contact:
    name: 蔓蔓点菜开发组
servers:
  - url: https://express-xxx.sh.run.tcloudbase.com
    description: 云托管生产环境（私有链路，仅 wx.cloud.callContainer 可访问）
  - url: http://127.0.0.1:8080
    description: 本地开发
paths:
...
  /api/v1/dishes:
    get:
      operationId: listDishes
      summary: 查询菜品列表
      security: []
      parameters:
        - name: category
          in: query
          schema:
            $ref: '#/components/schemas/DishCategory'
          description: 按分类过滤（hot | cold | soup | staple）
      responses:
        '200':
          description: 菜品列表
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Dish'
                  requestId:
                    type: string
              example:
                data:
                  - id: "dish-tomato-egg"
                    name: "鸡蛋西红柿"
                    category: "hot"
                    description: ""
                    imageUrl: ""
                    isActive: true
                    sortOrder: 10
                  - id: "dish-rice"
                    name: "米饭"
                    category: "staple"
                    description: ""
                    imageUrl: ""
                    isActive: true
                    sortOrder: 60
                requestId: "req-abc123"

    post:
      operationId: createDish
      summary: 创建菜品（用户菜品，关联当前用户）
      security:
        - AuthHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDishPayload'
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Dish'
                  requestId:
                    type: string
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/dishes/{id}:
    get:
      operationId: getDish
      summary: 按 ID 获取菜品
      security: []
      parameters:
        - $ref: '#/components/parameters/DishId'
      responses:
        '200':
          description: 菜品详情
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Dish'
                  requestId:
                    type: string
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      operationId: updateDish
      summary: 更新菜品（仅所有者可操作）
      security:
        - AuthHeader: []
      parameters:
        - $ref: '#/components/parameters/DishId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateDishPayload'
      responses:
        '200':
          description: 更新后的菜品
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Dish'
                  requestId:
                    type: string
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      operationId: deleteDish
      summary: 删除菜品（软删除，仅所有者可操作）
      security:
        - AuthHeader: []
      parameters:
        - $ref: '#/components/parameters/DishId'
      responses:
        '200':
          description: 删除成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    nullable: true
                  requestId:
                    type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # -------------------------------------------------------------------------
  # 点菜计划
  # -------------------------------------------------------------------------
...
  /api/v1/admin/dishes:
    get:
      operationId: adminListDishes
      summary: 管理员查询全部菜品（含已停用）
      security:
        - AdminHeader: []
      responses:
        '200':
          description: 全部菜品
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Dish'
                  requestId:
                    type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

    post:
      operationId: adminCreateDish
      summary: 管理员创建菜品
      security:
        - AdminHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminCreateDishPayload'
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Dish'
                  requestId:
                    type: string
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/admin/dishes/{id}:
    get:
      operationId: adminGetDish
      summary: 管理员按 ID 获取菜品
      security:
        - AdminHeader: []
      parameters:
        - $ref: '#/components/parameters/DishId'
      responses:
        '200':
          description: 菜品详情
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Dish'
                  requestId:
                    type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      operationId: adminUpdateDish
      summary: 管理员更新菜品
      description: |
        支持更新：name、category、description、imageUrl、isActive、sortOrder。
        imageUrl 必须为 `cloud://` 开头且不超过 512 字符。
      security:
        - AdminHeader: []
      parameters:
        - $ref: '#/components/parameters/DishId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminUpdateDishPayload'
      responses:
        '200':
          description: 更新后的菜品
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Dish'
                  requestId:
                    type: string
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/admin/notifications:
    get:
      operationId: adminListNotifications
      summary: 管理员查询通知任务列表
      security:
        - AdminHeader: []
      parameters:
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/NotificationStatus'
          description: 按状态过滤（pending | sent | failed | no_quota | rejected）
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
      responses:
        '200':
          description: 通知任务列表
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/NotificationJob'
                  requestId:
                    type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/admin/notifications/{id}/retry:
    post:
      operationId: adminRetryNotification
      summary: 管理员重试失败的通知任务
      description: |
        - 若任务状态为 `failed` 或 `no_quota`，扣除配额后重试；
        - 若状态为 `sent`，返回 `409 ALREADY_SENT`；
        - `in_app` 类型任务跳过配额扣减。
      security:
        - AdminHeader: []
      parameters:
        - $ref: '#/components/parameters/NotificationJobId'
      responses:
        '202':
          description: 已加入重试队列
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/NotificationJob'
                  requestId:
                    type: string
        '400':
          description: 任务状态不允许重试（如 pending）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: 任务已发送成功，不可重试
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/admin/subscriptions:
    post:
      operationId: adminCreateSubscription
      summary: 管理员为用户开通知额
      description: |
        管理员通过此接口为指定用户开通通知权限并授予额度。
        同一用户重复调用更新剩余额度，不重复创建记录。
      security:
        - AdminHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - openid
                - templateId
                - quota
              properties:
                openid:
                  type: string
                  description: 用户 OpenID
                templateId:
                  type: string
                  description: 订阅消息模板 ID
                quota:
                  type: integer
                  minimum: 0
                  description: 授予的订阅额度
      responses:
        '201':
          description: 创建/更新成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Subscription'
                  requestId:
                    type: string
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

  # -------------------------------------------------------------------------
  # 内部接口（notify-admin 云函数调用）
  # -------------------------------------------------------------------------
...
components:
  securitySchemes:
    AuthHeader:
      type: apiKey
      in: header
      name: X-WX-OPENID
      description: 微信用户 OpenID，由小程序 cloud.getWXContext() 获取并透传
    AdminHeader:
      type: apiKey
      in: header
      name: X-WX-OPENID
      description: 管理员 OpenID（需在 ADMIN_OPENIDS 环境变量中配置）
    InternalToken:
      type: apiKey
      in: header
      name: X-Notify-Token
      description: notify-admin 与 Express 服务共享的访问令牌（NOTIFY_API_TOKEN 环境变量）
  parameters:
    DishId:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: 菜品 ID
...
    NotificationJobId:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: 通知任务 ID
  schemas:
    DishCategory:
      type: string
      enum: [hot, cold, soup, staple]
      description: 菜品分类
    Dish:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
          maxLength: 30
        category:
          $ref: '#/components/schemas/DishCategory'
        description:
          type: string
          maxLength: 100
        imageUrl:
          type: string
          format: uri
        isActive:
          type: boolean
        sortOrder:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateDishPayload:
      type: object
      required: [name, category]
      properties:
        name:
          type: string
          maxLength: 30
        category:
          $ref: '#/components/schemas/DishCategory'
    UpdateDishPayload:
      type: object
      properties:
        name:
          type: string
          maxLength: 30
        category:
          $ref: '#/components/schemas/DishCategory'
    AdminCreateDishPayload:
      type: object
      required: [name, category]
      properties:
        name:
          type: string
          maxLength: 30
        category:
          $ref: '#/components/schemas/DishCategory'
        description:
          type: string
          maxLength: 100
        imageUrl:
          type: string
          maxLength: 512
          description: 必须为 cloud:// 开头（微信云存储）
        isActive:
          type: boolean
          default: true
        sortOrder:
          type: integer
          default: 0
    AdminUpdateDishPayload:
      type: object
      properties:
        name:
          type: string
          maxLength: 30
        category:
          $ref: '#/components/schemas/DishCategory'
        description:
          type: string
          maxLength: 100
        imageUrl:
          type: string
          maxLength: 512
        isActive:
          type: boolean
        sortOrder:
          type: integer
...
    NotificationStatus:
      type: string
      enum: [pending, sent, failed, no_quota, rejected]
    NotificationJob:
      type: object
      properties:
        id:
          type: string
        mealPlanId:
          type: string
        mealPlanVersion:
          type: integer
        channel:
          type: string
          enum: [in_app, wechat_subscribe]
        recipientOpenid:
          type: string
        status:
          $ref: '#/components/schemas/NotificationStatus'
        attemptCount:
          type: integer
        lastErrorCode:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        sentAt:
          type: string
          format: date-time
          nullable: true
    Subscription:
      type: object
      properties:
        openid:
          type: string
        templateId:
          type: string
        remainingQuota:
          type: integer
        hasSubscription:
          type: boolean
        acceptedAt:
          type: string
          format: date-time
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
          required: [code, message]
      required: [error]
      example:
        error:
          code: "VALIDATION_ERROR"
          message: "缺少必填字段"
  responses:
    BadRequest:
      description: 请求参数校验失败
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: "VALIDATION_ERROR"
              message: "至少选择一道菜"
    Unauthorized:
      description: 未认证（缺少 X-WX-OPENID）
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: "UNAUTHORIZED"
              message: "缺少用户身份"
    Forbidden:
      description: 无权限
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: "FORBIDDEN"
              message: "无权操作此资源"
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: "NOT_FOUND"
              message: "记录不存在"
...